+++ inc/LocationConfig.hpp
@@ -1,0 +1,39 @@
+#ifndef LOCATIONCONFIG_HPP
+#define LOCATIONCONFIG_HPP
+
+#include <string>
+#include <vector>
+#include <map>
+#include <utility>
+#include <netinet/in.h>
+#include <sys/socket.h>
+#include <iostream>
+#include "colors.hpp"
+#include <sstream>
+#include <cstdlib>
+#include <cerrno>
+#include <cstring>
+#include <sys/stat.h>
+#include <unistd.h>
+#include <netinet/in.h>
+#include <sys/socket.h>
+
+class LocationConfig
+{
+public:
+    std::string path;
+    std::vector<std::string> methods;
+    bool autoindex;
+    std::string index;
+    std::string root;
+    std::string upload_path;
+    std::string cgi_path;
+    std::vector<std::pair<int, std::string> > returns;
+
+    LocationConfig();
+    LocationConfig(const LocationConfig& other);
+    LocationConfig& operator=(const LocationConfig& other);
+    ~LocationConfig();
+};
+
+#endif
