+++ inc/colors.hpp
@@ -1,15 +1,3 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   colors.hpp                                         :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: syl <syl@student.42.fr>                    +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2025/01/27 09:18:35 by sforster          #+#    #+#             */
-/*   Updated: 2025/12/04 13:13:59 by syl              ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
 #ifndef COLORS_HPP
 # define COLORS_HPP
