+++ src/parsing/01_server_parsing.cpp
@@ -19,149 +19,6 @@
 #include <algorithm>
 #include <cstdlib>
 
-// ============================================================================
-// LocationConfig
-// ============================================================================
-
-LocationConfig::LocationConfig()
-	: path()
-	, methods()
-	, autoindex(false)
-	, index()
-	, root()
-	, upload_path()
-	, cgi_path()
-	, returns()
-{
-}
-
-LocationConfig::LocationConfig(const LocationConfig& other)
-	: path(other.path)
-	, methods(other.methods)
-	, autoindex(other.autoindex)
-	, index(other.index)
-	, root(other.root)
-	, upload_path(other.upload_path)
-	, cgi_path(other.cgi_path)
-	, returns(other.returns)
-{
-}
-
-LocationConfig& LocationConfig::operator=(const LocationConfig& other)
-{
-	if (this != &other)
-	{
-		path = other.path;
-		methods = other.methods;
-		autoindex = other.autoindex;
-		index = other.index;
-		root = other.root;
-		upload_path = other.upload_path;
-		cgi_path = other.cgi_path;
-		returns = other.returns;
-	}
-	return *this;
-}
-
-LocationConfig::~LocationConfig()
-{
-}
-
-// ============================================================================
-// ServerConfig
-// ============================================================================
-
-ServerConfig::ServerConfig()
-	: listen_port(0)
-	, server_name()
-	, error_pages()
-	, root()
-	, client_max_body_size(0)
-	, locations()
-	, fd_socket_serv(-1)
-	, sockaddr()
-	, server_len(0)
-	, error_code_message()
-	, mime_types()
-{
-}
-
-ServerConfig::ServerConfig(const ServerConfig& other)
-	: listen_port(other.listen_port)
-	, server_name(other.server_name)
-	, error_pages(other.error_pages)
-	, root(other.root)
-	, client_max_body_size(other.client_max_body_size)
-	, locations(other.locations)
-	, fd_socket_serv(other.fd_socket_serv)
-	, sockaddr(other.sockaddr)
-	, server_len(other.server_len)
-	, error_code_message(other.error_code_message)
-	, mime_types(other.mime_types)
-{
-}
-
-ServerConfig& ServerConfig::operator=(const ServerConfig& other)
-{
-	if (this != &other)
-	{
-		listen_port = other.listen_port;
-		server_name = other.server_name;
-		error_pages = other.error_pages;
-		root = other.root;
-		client_max_body_size = other.client_max_body_size;
-		locations = other.locations;
-		fd_socket_serv = other.fd_socket_serv;
-		sockaddr = other.sockaddr;
-		server_len = other.server_len;
-		error_code_message = other.error_code_message;
-		mime_types = other.mime_types;
-	}
-	return *this;
-}
-
-ServerConfig::~ServerConfig()
-{
-}
-
-/*
-ServerConfig WebServ::getServer(int index)
-{
-	if (index < 0 || static_cast<size_t>(index) >= servers.size())
-		throw std::out_of_range("Server index out of range");
-	return servers[index];
-}
-
-std::string toString(int value) {
-    std::ostringstream oss;
-    oss << value;
-    return oss.str();
-}
-
-//hostReq format: "hostname:port"
-ServerConfig WebServ::getServer(std::string hostReq)
-{
-	ServerConfig emptyServer;
-	emptyServer.listen_port = -1;
-	emptyServer.server_name = "";
-	//for each server create string that match the request format
-	std::string serverHostPort;
-	for (size_t i = 0; i < servers.size(); ++i)
-	{
-		serverHostPort = servers[i].server_name + ":" + toString(servers[i].listen_port);
-		if (serverHostPort == hostReq)
-			return servers[i];
-	}
-	for (size_t i = 0; i < servers.size(); ++i)
-	{
-		//if no exact match found, try to match only by port
-		serverHostPort = ":" + toString(servers[i].listen_port);
-		if (serverHostPort == hostReq)
-			return servers[i];
-	}
-	return emptyServer;//return first server as default
-
-}*/
 
 void WebServ::addServer(ServerConfig config)
 {
@@ -298,7 +155,6 @@ void WebServ::parseConfig(std::string path)
 						std::stringstream sToken(tokens[i]);
 						int code;
 						sToken >> code;
-						//int = std::atoi(tokens[i].c_str());
 						currentServer.error_pages[code] = pathErr;
 					}
 				}
