--- src/parsing/parsing_main.cpp
@@ -1,7 +1,0 @@
-#include "Webserv.hpp"
-
-void WebServ::defaultConfig()  
-{}
-
-void WebServ::fileConfig()
-{}
