--- temp/GitHubTemplates/template_branch_commit.md
@@ -1,27 +1,0 @@
-# Commit Template
-## On all branches (except main and dev)
-
-main branch(stable), final project<br>
-dev branch(unstable and collaborative), default branch<br>
-other branches(personnal) based on dev. A branch per issue/ task / debug 
-
-## Importance:
-Personnal
-
-## Goal:
-To register your own work (at home/ at school)
-
-## Type of commit: 
-Simple commit on the branch<br>
-
-## When:
-Everyday or when changing computer from home to school<br>
-
-## Name:
-Not important, only for you
-
-## Description:
-Not important, only for you
-
-## Protocol
-Not important, only for you<br>
