--- temp/GitHubTemplates/template_dev_pull_request.md
@@ -1,37 +1,0 @@
-# Pull Request Template
-## From all branches to dev (except main)
-
-main branch(stable), final project<br>
-dev branch(unstable and collaborative), default branch<br>
-other branches(personnal) based on dev. A branch per issue/ task / debug 
-
-## Importance:
-Serious stuff to do quite often
-
-## Goal:
-To add your work (issues) on the collaborative project<br>
-To review team's member work<br>
-To keep a workflow with clear tasks(issues) and history
-
-## Type of Pull Request: 
-Squash and merge commit (erase history of your own commits)<br>
-From your branch/issue/feature(personnal) to dev(collaborative)
-
-## When:
-When you are done with an issue<br>
-when you want to discuss/ test your work on the dev branch
-
-## Name:
-Task name (probably the issue's name?)
-
-## Description:
-Description of the issue<br>
-Maybe some comments / doubts / ideas<br>
-Etc...
-
-## Protocol
-1 Pull request on github with description and name<br>
-2 Tell others<br>
-3 Review + changes (add comment on github)<br>
-4 When everything ok merge<br>
-5 Delete the branch on github (next task will be an other branch)<br>
