--- temp/GitHubTemplates/template_main_pull_request.md
@@ -1,25 +1,0 @@
-# Pull Request Template
-## From dev(unstable) to main(stable)
-
-## Importance:
-Serious stuff
-
-## Goal:
-To keep track of a stable version of the project
-
-## Type of Pull Request: 
-Merge commit (keep history of merge commits on dev branch) from dev(unstable) to main(stable) ONLY!
-
-## When:
-When features have been tested on the dev branch and are stable<br>
-Before some changes may break the dev branch
-
-## Name:
-A brief description
-
-## Description:
-A list of all changes made on the dev branch<br>
-Steps before, state of the project<br>
-A reflexion on next steps to do<br>
-If an upcoming change may break the dev branch, please decribe this change<br>
-Etc...
