--- temp/doc/theorytoknow.txt
@@ -1,12 +1,0 @@
-telnet 
-ssh
-
-TCP/IP
-
-RFC 791
-
-TCP stream socket (reliable two-way connected communication streams.)
-UDP Datagram socket (video games for exemple)
-
-what is a port?
-what is a socket?
