--- temp/doc/webser eval sheet
@@ -1,15 +1,0 @@
-<?xml version="1.0" standalone="no"?>
-<xournal creator="xournalpp 1.2.7" fileversion="4">
-<title>Xournal++ document - see https://xournalpp.github.io/</title>
-<preview>iVBORw0KGgoAAAANSUhEUgAAAGIAAACACAIAAAB/SyvQAAAABmJLR0QA/wD/AP+gvaeTAAAQGklEQVR4nO1d23IbR5LN6q6+X9AgSJGSPLbG9nge5v//YscR++CN3QiHRtaFAgni1uhbdXVX7cMBa2BKpCnbQ1BqnAcFBOJSXajKyjx5MptpremA3wJ//fp10zR1XTPGfvjhB9/39z2kxwg2nU4tyyKiuq6jKCrLMoqiruuklJZlnZ2d4a8DB+u6ru9727aVUpZlaa0ZY0SktTaPAaUUY8yyLKWU4zj7G/MewBeLxZs3b8IwxBxZltX3veM46/U6jmM8CMNQKaW1xgSNx+OTk5N9j/xBwbC/HMdp29Z1XawpIur73rIsrB2lFBFprTnnWGK2be975A8Kvlgs5vM5Y8z3/aIofN+v61prjSnD1MBOua4rhDg5OfF9P0mSfY/8QcEuLi76vmeMSSk559h0nueVZckYe/78+cGEEzYdbDNjDPNFRNhrMFIwSWayjFG3LGvXwH/Z4FdXV1dXV1LKIAgsy+q6DsdZ0zREBDeKMda2red5Sqm2bSeTSRAEcRxzzvc9/gcCJ6LRaASfwLZtLKi+79M0xaQ4jqOU4pz7vq+17vseRn04c0RE3HVdzjnnvOu69Xp9cnKitW6axvd9KaXrurZt13WdZZmUUkoZhiEOwX2P/EHBzs/PLy4utNZnZ2dPnz7d93geKVjf913XwbeEL24eEJFSyrZthMeMMbjmePGwNh1MOE660Wjk+z6sOGPMdV14CUqppmls2/Y8D8GNUurJkyfDOemYIVJgvIdmdO4J/ssvvywWC8/zmqZJkgQLBz5RFEXv37/HwQ9fQUoZRVHTNKenp0dHR8MJgNlsNmvbtu/7LMsQguw6kwDchT2N8FGA9X3f9/32P4zhX+OXm9fBopsnh7Y9+XK5PD8/R2Tb973W2vf9siw555gvBMOc8zAMgyAgItu2OedxHO978A8H1nUdEcGTxJKBGQIxgGMOKwieN/47NGaOr9fry8tLIQTCNyllkiRN0yAkHo1GVVUR0WazOTs7C4LAcZy6rj3PG4/H+x78w4HtZlbgE8CNNOScMd43ON9BgV9eXr5+/TpNUyml7/sXFxej0ahtW9/38zxP07Rt2yAI8jz3fd91XcdxkiThnK/X62+++Wbf438gsBt5OmOe8N/d1fShozAc8M1mUxQFSLi2bW3bjuO4bVvsO9gsxhjnHOcg57wsy6Ojo0HxvOzly5egwEF1a60dx4FDAIe7KIrnz5/jgAPlkmXZvof90GCr1QrRSV3XrusivwJHyXEckAfgCaSUcAWOjo48z3McZzgWnTdN8+7dO/iKy+XScZwgCJqm2Ww2x8fHUkrsx9FolOc5aCnLslar1WQyGU4EszXhhlH68BVDttwG/PXr19PpNI5jbC4pZd/3MNhRFGmt8zyPoghZFiKyLGu5XGZZ9uLFC8/z9j3+BwKbTqfGZjPGkP7F4oI9sm2bMXZ2drbvoe4TLM/zPM+bphmPx1VVmewbHKi+713X1VpLKYUQURQRkeu6x8fHex74w4K9f/9+uVx6nldVFSQCmCkhRJIkQgjY6aZpXNcloq7rJpPJoDg5upH1petFRETICxiNCuZuV+IzKJKAT6fTzWYDqUVZllrrJEmWyyU8ScwdcsJZlsGBQsqAcz4ejwfiOrGiKKqqatt2VxMQhuF6vYZVsm0bnBRiYMYYUi+np6cDmSPCamqaBvkC3/cdx0GM4nmemQXoeLquc123KIrj4+OhSTRZXdfgKumQGrgFruvyruts226aBmHKweH+KHjbtkIIpVTXdU+ePEFS4IAb4HmezWYqCNiTJ1YQDMUkfyqY1rrvSWsaknLik8FfvVJE9OLFwSTdBfbuXV8U2nWZ41Db0l//epivj4CnKQsCJqW2LBaGpBQdzroPwRcLPZ2q0Yj9/e+H6bkVFueUZcz32b/+pfY9mMcLJoTuOi0lOQ6F4cEh+Dj4y5f/C9n3aDRaLpfffffdoKQm9wQrisIov0EwERHYkkGlmO4Gb9t2sVigqAmlBF3XIf17fHwMVvcAPh6PAx7Mf5xzh0shNdee7bVl64c+MaLDLBERgpWyLEFgoo4AqRRD7CIZZVkWEi1pmoKcM5zvEASGnIiUUnVdI7sLohI8N+e8qqo4jqEEA40JFR3KonzfX6/XRtv6BYOhQmW5XBoNGJYJVhmuH6ZKKQUhHbhwy7I8z0P605h8zG9d17Zth2GYpqlJPdB1/ZlRmpknzWj0NfCax1PIwIkIBfZCCCJijIGiQwlP3/dFUXie1/f9ZrMRQiDliT9hykzaqmkaIUTXdScnJ/P5XCn15s2bKIrquk7TdLPZhGGIr8Mk9n0PGj6OYyEE8jT4kdq2JaIsyzCzvu9zzoUQlmWFYZjnOefctu0oivI8J6K+75Mk2Ww2QRCAOzMJIVRRNk0D44BfBTvj9PT0ntPEtNZCiDzP8Ts3TbOrRTGypq7rwjBcrVZhGKJcCoVloPRQnWgEUGEYGufrbiniHxQomFQY/XpVmo/dLcHZrYv/VDCt9Wq1ms1mqJLDNOGng4SwaZqyLI+Pj/ErwTztFkmXZZmmKerMwjAMwxBZYsdxIGiBRBFOLCYXTGmaplhfjx/baYIADFcSBIEQArWY2AtYTdh6mIKzs7OiKIgISijzq+IFyPrVde04jqk9N3/1fR/VjI7jCCGCINBa13X91VdfPVpvlhMRhovlgxlBAqqqKhQf2rYNnRjaXti2fXV1RUSwYpgjKSUReZ4HOwVPFbJEmKHPunUB01oXRTGfz9frdZqmZVkqpdC9AbYWYkOz/4MggLw3SZL1eh1FEdyFMAyllEdHR/u+ov8IGCQ7S+FEVh1Hv20phqkK226WRdl3XXefJkUDnCMissqy/Omnn/z6zfT9uSmFOuAGGCIPui4CM744XRf64km6Lu2BmPfx+McPA355eQmntuu64+NjwzrBdYTCAjEw5qvruiiKhkbdbZW8cB3xFHwZuhb27orBUEVuHpjozDyg6zVo5GT4k3Gs8Bh/Auvw8Nf8O8CJCF64iXgty4LzHQQBuE3EcWVZwhsw9eNg8qD1KYoCajHMGtYggilEMygVTpIkz3N46q7rfi6035Yh2Gw2cIWhba6qKkkSxhhifawFLASl1BD1TVprhP7QORubjW2olELsglkryxJRGII1rTW4FATrV1dXsGVBEKxWK3RZ2a3i4Jy7rovQD+9SSmVZ9vgjuy17+e7dOyKq6zpJEmwo3/ch45VSNk2DFaSUGpSW0ODfRIpxAnDAgS1ACSIRwdZCVwihJng4GCxMK2JdhM1SymfPnn0xE8qm02kURf+zCP7izs9OT76YC/tzcbM684CPgr969cp13cViMZlMDo2JbgMDaS2lhMIZfDPiYbRTBakEPtNxnM+aNvrd4FprmOT1ej0ajQwFDu95aCU8t2HrXq7Xa5SImcAN7nKWZZvNBhzmkLekRdeNCOu6rusaIYjjOJvNBlk5NJ9jjIGlHCa2JK8QoqoqENtEhIAOr0AlHRxldFr9+uuvh+Y3QPDcLxYLowZAGGEK68AQdF2XZdlwKsNugL99+zbLsv9ejP8xWhzcy9vAGWPT6fRbx0qTIQZr98TBC78X+Hw+h0uJOqgwDMEQPHv2bN9je0TY+k11XaMhWJ7niPiNMAGinCiKPhem8T+BLd+0Xq8hMjESVSPkQO5fKRXHMZqImxrFyWSCcnKj9DAlrnTNgqNnDzg/49x/domZLRfeNI1p5ozZqarKdCk2ekPoT4hoNBqVZVnX9WazieM4CAIpJcQ96/U6SRI0pgFvhcSyaQSBnj34Fixh/FRhGMJBQ/AEZn2z2aB3hpl0iDVAn4JrxueAeu/7HkNN0xSt30wfBc/zUIQ7Ho8/lS9l4NKMTM34SvCh8K3g5LABsejM+7Elb2ustntnhA/bZd2hNrrtTx/KlG680vSCvVGQ+2GLyk8CQ4xC152JjXaNiCCzpGsxIL7J931IMPD8aDT6fV/8eYGhnew/L/wfooN7eSuYycQZwehnZ18fANsqAyg+Oeer1erbb7/d96geHbaZlbIsoczFDQuICFpdHFVd18VxDCkuTD6YFnhY8B6+bGEBQ7J3NptB92bUKSAMqqqKogiH3enp6WDrfZjWGjLe6XSapikE4rjpk+u6lmXBxzFNxOM4Hk4TMAOmtV4ul0QEuQB6YH7//fcDXDJ3YFtlwDmfz+dEBK3ILnt5AEEtxzkXnXYsItJ/xFX9gmHNZrOiKP7rTX9xOTvM0W3gfd/PZrOvLGucHdjLW3FgL+8FTkRlXm5ebZRSohWWZfmeLzvpOq5oBV1rKYUQcRQrpXrV03Vc3vd94Ad1U7uOS0Syk4axCvxASlk3ted6ohUo57Atu5VtGIRVXeG/mjQj1soWsbTv+et8/eIvL5RWtrXN7mwr7CJmj/cm1GRa6zIv3//fe6VUVVcYq5EvQerMrmtaUZhhGGFzlyOQRxC54rFoxVF2tFgtGGNpkppbuwZ+UNUVqoHKqvQ9Pwqj5XqZxAkIvKZpIMsEFwg5lWjF0emRP/Hx+SZagFDtAZRp29D3qnbGXhv43oeVaHe8+e4ma3eUbXyUTjJqX/Ou+4zhYbAlA0TXCxK+55oh3qfo4u4LuOMTPvrj4/W773o8hR98Nps1TRPH8VyIsthMJpOhqXTvg+1JB9HuLleLBwiA9za6RwNORJeXl2VZLpdLFMqZe9UxxoIgsG07CIKyLBEJV1UVBAFyB2DjjaoZZIvneUKI58+f7/vS/kxwIkrTVGudZVlZluPxGOlMtOrHYSelRL0uKsch9KHrBgbIgOI1KOKcTCb7vq4/Gdta36IosLm6rsOd/JBBwc1XpJSnp6emppd27khu0jB4Ep4EEX2Oybg7wLTWeZ6fn5/btg2JE5zJKIqQFEPtPlwVNLmADwUlFOe8bVu4Tlh3SLRJKXErBDCcQojRaAThHaqosSTDMCzLEhUgcJpQk20KYpBZy7KsbVtI1eM4RhEMdr2p1CqKIkkSrfUNdSQyhrs5QVgVFHnjGoMgQA4RXRMgbcdnxnE8Ho+3Jvzi4gKV82BvoZCDVA5sLzKUJhuM98PVVEqhEhpXiNwh0oe+76dpuu+l8CeAvX37djQa/XgZ/C08yOdvBTNV4QfcBtd1+c8//4xbY4ZheBDr3AaGUxzSeCMg2K34pV8nx03i0zwgIhisXRUCXaft1fXdJHdz/wjWPmxaYgRD+vrOb+YkNWfrbnHfg2lBGW6VtatFQbkvTiIUFyDR0jQN7ryCAxFjxVWh+YmR6aBBClwqnFkw+VCDmHZRpiMGxCr46q7r8OE4bnBKIgmGHwwHX1VVnuc92N2o2Pn5OW6VhYv3ff+w9T7Egb28F/hqtYIlwn1WTNsqz/OKooC+C8+YiucvLxb5TXAUrv544f8t6k5/7TcN6iaid4NVVYW2OvseyaOGJYTAIXLAHeBEJISYzWZpmsI8EZFpT4WA8OnTpwMPYrjWOo7jq34chO2X2qTqj8OCI+taPSiFfY/nkeL/AX/0VKIpXOSaAAAAAElFTkSuQmCC</preview>
-<page width="1878.96" height="2466">
-<background type="pdf" domain="absolute" filename="/home/syl/42/WebServ/webser eval sheet.pdf" pageno="1"/>
-<layer>
-<stroke tool="highlighter" color="#ff00ff7f" width="13.289062" fill="60" capStyle="butt">35.999999 1012.9511 183.19335 1012.9511</stroke>
-<stroke tool="highlighter" color="#ff00ff7f" width="0" fill="60" capStyle="butt">183.19335 1019.5957 183.19335 1019.5957</stroke>
-<stroke tool="highlighter" color="#ff00ff7f" width="0" fill="60" capStyle="butt">948.49215 1996.0956 948.49215 1996.0956</stroke>
-<stroke tool="highlighter" color="#ff00ff7f" width="13.289062" fill="60" capStyle="butt">35.999999 2002.9511 1090.0078 2002.9511</stroke>
-<stroke tool="highlighter" color="#ff00ff7f" width="0" fill="60" capStyle="butt">1090.0078 2009.5956 1090.0078 2009.5956</stroke>
-</layer>
-</page>
-</xournal>
