--- temp/tester_v2/docker/Dockerfile
@@ -1,26 +1,0 @@
-# Use Ubuntu base image
-FROM ubuntu:22.04
-
-# Prevent interactive prompts during install
-ENV DEBIAN_FRONTEND=noninteractive
-
-# Install nginx
-RUN apt-get update && \
-    apt-get install -y nginx && \
-    apt-get clean && \
-    rm -rf /var/lib/apt/lists/*
-
-# Create directories for our test files
-RUN mkdir -p /var/www/html \
-             /var/www/cgi \
-             /var/www/errors \
-             /var/www/uploads
-
-# Copy our custom nginx config
-COPY nginx.conf /etc/nginx/nginx.conf
-
-# Expose port 80
-EXPOSE 80
-
-# Start nginx in foreground
-CMD ["nginx", "-g", "daemon off;"]
