--- temp/tester_v2/docker/nginx.conf
@@ -1,57 +1,0 @@
-# Force nginx to use HTTP/1.0
-events {
-    worker_connections 1024;
-}
-
-http {
-    # Include MIME types (so nginx knows .html = text/html, .jpg = image/jpeg, etc.)
-    include /etc/nginx/mime.types;
-    default_type application/octet-stream;
-
-    # Force HTTP/1.0 for proxy (if needed)
-    proxy_http_version 1.0;
-
-    server {
-        # Listen on port 80 inside container
-        listen 80;
-        server_name localhost;
-
-        # Root directory for HTML files
-        root /var/www/html;
-        index index.html;
-
-        # Main location - serve HTML files
-        location / {
-            # Only allow GET and HEAD methods
-            limit_except GET HEAD {
-                deny all;
-            }
-        }
-
-        # Uploads - Allow POST and DELETE
-        location /uploads/ {
-            limit_except GET POST DELETE {
-                deny all;
-            }
-            autoindex on;
-        }
-
-        # CGI location
-        location /cgi-bin/ {
-            alias /var/www/cgi/;
-            # Only allow GET and POST
-            limit_except GET POST {
-                deny all;
-            }
-        }
-
-        # Error pages
-        location /errors/ {
-            alias /var/www/errors/;
-        }
-
-        # Custom error pages
-        error_page 404 /errors/404.html;
-        error_page 405 /errors/405.html;
-    }
-}
