--- temp/tester_v2/fixtures/errors/400.html
@@ -1,8 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head><title>400 Bad Request</title></head>
-<body>
-<h1>400 Bad Request</h1>
-<p>The request could not be understood by the server.</p>
-</body>
-</html>
