--- temp/tester_v2/fixtures/errors/403.html
@@ -1,8 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head><title>403 Forbidden</title></head>
-<body>
-<h1>403 Forbidden</h1>
-<p>Access to this resource is forbidden.</p>
-</body>
-</html>
