--- temp/tester_v2/fixtures/errors/403_forbidden.html
@@ -1,8 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head><title>403 Forbidden</title></head>
-<body>
-<h1>403 Forbidden</h1>
-<p>You do not have permission to access this resource.</p>
-</body>
-</html>
