--- temp/tester_v2/fixtures/errors/404.html
@@ -1,8 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head><title>404 Not Found</title></head>
-<body>
-<h1>404 Not Found</h1>
-<p>The requested resource was not found on this server.</p>
-</body>
-</html>
