--- temp/tester_v2/fixtures/errors/405.html
@@ -1,8 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head><title>405 Method Not Allowed</title></head>
-<body>
-<h1>405 Method Not Allowed</h1>
-<p>The requested method is not allowed for this resource.</p>
-</body>
-</html>
