--- temp/tester_v2/fixtures/errors/500.html
@@ -1,8 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head><title>500 Internal Server Error</title></head>
-<body>
-<h1>500 Internal Server Error</h1>
-<p>The server encountered an internal error and could not complete your request.</p>
-</body>
-</html>
