--- temp/tester_v2/fixtures/index.html
@@ -1,8 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head><title>Index</title></head>
-<body>
-<h1>Welcome to Webserv</h1>
-<p>This is the index page.</p>
-</body>
-</html>
