--- temp/tester_v2/fixtures/tiny.html
@@ -1,1 +1,0 @@
-<h1>Hello from nginx!</h1>
